refactor: expose auth runtime through public API
